Admin: Refactor components
